+++ navigation/src/commonMain/kotlin/com/hoc081098/solivagant/navigation/internal/StackEntry.kt
@@ -37,7 +37,6 @@ import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import com.hoc081098.solivagant.lifecycle.Lifecycle
-import com.hoc081098.solivagant.lifecycle.LifecycleOwner
 import com.hoc081098.solivagant.navigation.BaseRoute
 import com.hoc081098.solivagant.navigation.ContentDestination
 import com.hoc081098.solivagant.navigation.NavRoot
@@ -50,7 +49,7 @@ internal class StackEntry<T : BaseRoute> private constructor(
   val id: StackEntryId,
   val route: T,
   val destination: ContentDestination<T>,
-  private val entryLifecycleOwner: StackEntryLifecycleOwner,
+  val lifecycleOwner: StackEntryLifecycleOwner,
 ) {
   //region Removed from backstack state
   private val isRemovedFromBackstackState: MutableState<Boolean> = mutableStateOf(false)
@@ -76,26 +75,24 @@ internal class StackEntry<T : BaseRoute> private constructor(
 
   //region Lifecycle
   internal fun moveToCreated() {
-    entryLifecycleOwner.maxLifecycle = Lifecycle.State.CREATED
+    lifecycleOwner.maxLifecycle = Lifecycle.State.CREATED
   }
 
   internal fun moveToStarted() {
     check(!isRemovedFromBackstack.value) { "Can not move to STARTED state if removed from backstack" }
-    entryLifecycleOwner.maxLifecycle = Lifecycle.State.STARTED
+    lifecycleOwner.maxLifecycle = Lifecycle.State.STARTED
   }
 
   internal fun moveToResumed() {
     check(!isRemovedFromBackstack.value) { "Can not move to RESUMED state if removed from backstack" }
-    entryLifecycleOwner.maxLifecycle = Lifecycle.State.RESUMED
+    lifecycleOwner.maxLifecycle = Lifecycle.State.RESUMED
   }
 
   internal fun moveToDestroyed() {
-    entryLifecycleOwner.maxLifecycle = Lifecycle.State.DESTROYED
+    lifecycleOwner.maxLifecycle = Lifecycle.State.DESTROYED
   }
 
-  internal val lifecycleOwner: LifecycleOwner get() = entryLifecycleOwner
-
-  internal fun handleLifecycleEvent(event: Lifecycle.Event) = entryLifecycleOwner.handleLifecycleEvent(event)
+  internal fun handleLifecycleEvent(event: Lifecycle.Event) = lifecycleOwner.handleLifecycleEvent(event)
   //endregion
 
   companion object {
@@ -116,7 +113,7 @@ internal class StackEntry<T : BaseRoute> private constructor(
         id = StackEntryId(idGenerator()),
         route = route,
         destination = destination,
-        entryLifecycleOwner = StackEntryLifecycleOwner(
+        lifecycleOwner = StackEntryLifecycleOwner(
           hostLifecycleState = hostLifecycleState,
         ),
       )