+++ samples/solivagant-wasm-sample/src/wasmJsMain/kotlin/com/hoc081098/solivagant/sample/wasm/start/StartScreenRoute.kt
@@ -1,0 +1,12 @@
+package com.hoc081098.solivagant.sample.wasm.start
+
+import com.hoc081098.solivagant.navigation.NavRoot
+import com.hoc081098.solivagant.navigation.ScreenDestination
+
+data object StartScreenRoute : NavRoot
+
+val StartScreenDestination = ScreenDestination<StartScreenRoute> { _, modifier ->
+  StartScreen(
+    modifier = modifier,
+  )
+}