+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 ## [Unreleased] - TBD
 
+- [JetBrains Compose Multiplatform `1.6.0`](https://github.com/JetBrains/compose-multiplatform/releases/tag/v1.6.0).
+- [KMP ViewModel `0.7.1`](https://github.com/hoc081098/kmp-viewmodel/releases/tag/0.7.1).
+
+### Added
+
+- **New**: Add support for Kotlin/Wasm (`wasmJs` target) 🎉.
+- **New**: Add `LenientLifecycleRegistry`, a `LifecycleRegistry` without the state checking,
+  and has `moveTo` method to move to a specific state. This is a flexible version of `LifecycleRegistry`.
+- **New**: Add `rememberWindowLifecycleOwner()` for Desktop (JVM) platform.
+- **New**: Add `SavedStateSupport.ProvideCompositionLocals` and `SavedStateSupport.ClearOnDispose`.
+
+### Fixed
+- **Fixed**: an issue where `LocalLifecycleOwner` provided the wrong `LifecycleOwner` to the content of `NavHost` composable.
+
 ## [0.1.1] - Feb 25, 2024
 
 ### Fixed